refactor: 쿠폰 작성 폼 UI 컴포넌트 분리
